bluetooth: host: move GAP functionality to gap.h
Move GAP functionality from bluetooth.h to gap.h where it belongs to.
